refactor: skeletones for new layout
