+++ src/components/skeletones/episodeSkeleton/episodeSkeleton.module.css
@@ -1,0 +1,87 @@
+.container {
+  flex: 1 1 60%;
+  height: 12.3em;
+  container: mockCard / inline-size;
+}
+
+.mockСard {
+  --border-size: 1px;
+  --font-size-title: 1.2em;
+  --header-color-background: rgb(var(--input-color-background));
+
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+  gap: 0.5em;
+  padding: 0.5em;
+
+  border-block-start: var(--border-size) solid rgb(var(--border-color-light));
+  border-inline-start: var(--border-size) solid rgb(var(--border-color-light));
+  border-block-end: var(--border-size) solid rgb(var(--border-color-dark));
+  border-inline-end: var(--border-size) solid rgb(var(--border-color-dark));
+  border-radius: var(--border-radius-input);
+  box-shadow: 2px 2px 2px rgba(var(--shadow-color));
+}
+
+.mockСontentContainer {
+  padding: 0.5em;
+}
+
+.mockСardHeader {
+  height: 2em;
+  line-height: 2em;
+
+  font-weight: 700;
+  font-family: var(--font-title);
+  font-size: var(--font-size-title);
+  color: rgb(var(--background-color));
+  -webkit-line-clamp: 1;
+  line-clamp: 1;
+
+  border-radius: var(--border-radius-input);
+  background-color: var(--header-color-background);
+  border-block-end: var(--border-size) solid rgb(var(--border-color-light));
+}
+
+.mockСontent {
+  opacity: 0.2;
+  height: 1em;
+  margin-block-end: 0.5em;
+  border-radius: var(--border-radius-input);
+  background-color: rgb(var(--border-color-light));
+
+  &:last-child {
+    position: relative;
+    height: 2em;
+    margin-block-start: 1.65em;
+
+    &::after {
+      content: 'Close';
+      display: block;
+      position: absolute;
+      width: 100%;
+      line-height: 2em;
+      border-radius: var(--border-radius-input);
+    }
+  }
+}
+
+.animation {
+  opacity: 0.5;
+  background: linear-gradient(325deg, white, rgb(var(--border-color-dark)));
+  background-size: 400% 400%;
+  animation: skeletonAnimation 2s infinite;
+  animation-timing-function: cubic-bezier(0.25, 0.46, 0.45, 0.94);
+}
+
+@keyframes skeletonAnimation {
+  0% {
+    background-position: 1% 0%;
+  }
+  50% {
+    background-position: 84% 100%;
+  }
+  100% {
+    background-position: 1% 0%;
+  }
+}
