+++ src/components/skeletones/episodesSkeleton/episodesSkeleton.module.css
@@ -1,77 +1,109 @@
-.container {
-  flex: 1 1 60%;
-  height: 12.3em;
-  container: mockCard / inline-size;
-}
-
-.mockСard {
-  --border-size: 1px;
-  --font-size-title: 1.2em;
-  --header-color-background: rgb(var(--input-color-background));
+.skeleton {
+  --font-size-title: 1.3em;
 
   display: flex;
-  height: 100%;
+  flex: 1 1 auto;
+  max-height: 100%;
+  width: 100%;
+  opacity: 0.7;
+
+  position: absolute;
+}
+
+.skeletonLfet {
+  display: flex;
   flex-direction: column;
   gap: 0.5em;
-  padding: 0.5em;
+  flex: 1 1 50%;
 
-  border-block-start: var(--border-size) solid rgb(var(--border-color-light));
-  border-inline-start: var(--border-size) solid rgb(var(--border-color-light));
-  border-block-end: var(--border-size) solid rgb(var(--border-color-dark));
-  border-inline-end: var(--border-size) solid rgb(var(--border-color-dark));
-  border-radius: var(--border-radius-input);
-  box-shadow: 2px 2px 2px rgba(var(--shadow-color));
-}
-
-.mockСontentContainer {
-  padding: 0.5em;
-}
-
-.mockСardHeader {
-  height: 2em;
-  line-height: 2em;
-
-  font-weight: 700;
-  font-family: var(--font-title);
-  font-size: var(--font-size-title);
-  color: rgb(var(--background-color));
-  -webkit-line-clamp: 1;
-  line-clamp: 1;
+  block-size: 100%;
+  min-inline-size: 10em;
+  margin-block-end: 1em;
+  overflow-y: auto;
+  overflow-x: hidden;
 
   border-radius: var(--border-radius-input);
-  background-color: var(--header-color-background);
-  border-block-end: var(--border-size) solid rgb(var(--border-color-light));
-}
+  background-color: rgb(var(--input-color-background));
 
-.mockСontent {
-  opacity: 0.2;
-  height: 1em;
-  margin-block-end: 0.5em;
-  border-radius: var(--border-radius-input);
-  background-color: rgb(var(--border-color-light));
-
-  &:last-child {
-    position: relative;
-    height: 2em;
-    margin-block-start: 1.65em;
-
-    &::after {
-      content: 'Close';
-      display: block;
-      position: absolute;
-      width: 100%;
-      line-height: 2em;
-      border-radius: var(--border-radius-input);
-    }
+  &::-webkit-scrollbar {
+    width: 7px;
+    height: 5px;
+  }
+  &::-webkit-scrollbar-button {
+    width: 0px;
+    height: 0px;
+  }
+  &::-webkit-scrollbar-thumb {
+    background: var(--color-scroll);
+    border-radius: 0px;
+  }
+  &::-webkit-scrollbar-thumb:hover {
+    background: color-mix(in srgb, var(--color-scroll) 90%, white 10%);
+  }
+  &::-webkit-scrollbar-thumb:active {
+    background: color-mix(in srgb, var(--color-scroll) 90%, white 10%);
+  }
+  &::-webkit-scrollbar-track {
+    background: var(--color-scroll-background);
+    border-radius: 5px;
+  }
+  &::-webkit-scrollbar-track:hover {
+    background: var(--color-scroll-background);
+  }
+  &::-webkit-scrollbar-track:active {
+    background: var(--color-scroll-background);
+  }
+  &::-webkit-scrollbar-corner {
+    background: transparent;
   }
 }
 
-.animation {
-  opacity: 0.5;
-  background: linear-gradient(325deg, white, rgb(var(--border-color-dark)));
+.skeletonItem {
+  position: relative;
+  flex: 0 0 2.8em;
+  background: linear-gradient(to right, rgba(var(--input-color), 0.1) 40%, rgba(var(--input-color), 0.02) 100%);
+  border-radius: var(--border-radius-input);
+
+  animation: skeletonAnimation 1s infinite;
   background-size: 400% 400%;
-  animation: skeletonAnimation 2s infinite;
   animation-timing-function: cubic-bezier(0.25, 0.46, 0.45, 0.94);
+
+  &::before {
+    content: '';
+    position: absolute;
+    display: block;
+    width: 0.96em;
+    top: 50%;
+    left: 0.94em;
+    transform: translateY(-10%);
+    aspect-ratio: 1/1;
+    border: 1px solid rgb(var(--border-color-dark));
+    border-radius: 3px;
+    background-color: rgb(var(--background-color));
+  }
+}
+
+.skeletonRight {
+  flex: 1 1 50%;
+
+  padding-block-start: 1em;
+  padding-inline-start: 1.75em;
+  font-weight: 700;
+  font-family: var(--font-common);
+  font-size: 1.3em;
+  text-align: start;
+  color: rgb(var(--card-color));
+
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  line-clamp: 2;
+  -webkit-box-orient: vertical;
+
+  opacity: 0.2;
+  hyphens: auto;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  overflow-wrap: break-word;
 }
 
 @keyframes skeletonAnimation {
